pyproject.toml
[project]
name = "praise-announcement-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "pillow>=11.2.1",
]
